refactoring: устранены циклические ссылки
