+++ babel.config.js
@@ -1,0 +1,14 @@
+const presets = [
+  ['@babel/preset-env', {
+    targets: {
+      ie: '11',
+      firefox: '50',
+      chrome: '64',
+      safari: '11.1'
+    },
+
+    useBuiltIns: "entry"
+  }]
+];
+
+module.exports = { presets };
