+++ package.json
@@ -1,0 +1,37 @@
+{
+  "name": "mesto_russia",
+  "version": "0.9.0",
+  "description": "web page with user photos",
+  "main": "index.js",
+  "scripts": {
+    "build": "webpack --mode production",
+    "dev": "webpack serve"
+  },
+  "repository": {
+    "type": "git",
+    "url": "git+https://github.com/BobrikAU/mesto-project.git"
+  },
+  "author": "Bobrik Aleksandr",
+  "license": "ISC",
+  "bugs": {
+    "url": "https://github.com/BobrikAU/mesto-project/issues"
+  },
+  "homepage": "https://github.com/BobrikAU/mesto-project#readme",
+  "devDependencies": {
+    "@babel/core": "^7.17.9",
+    "@babel/preset-env": "^7.16.11",
+    "autoprefixer": "^10.4.5",
+    "babel-loader": "^8.2.5",
+    "clean-webpack-plugin": "^4.0.0",
+    "css-loader": "^6.7.1",
+    "cssnano": "^5.1.7",
+    "html-webpack-plugin": "^5.5.0",
+    "mini-css-extract-plugin": "^2.6.0",
+    "postcss-loader": "^6.2.1",
+    "webpack-cli": "^4.9.2",
+    "webpack-dev-server": "^4.8.1"
+  },
+  "dependencies": {
+    "core-js": "^3.22.3"
+  }
+}
