+++ postcss.config.js
@@ -1,0 +1,9 @@
+const autoprefixer = require('autoprefixer');
+const cssnano = require('cssnano');
+
+module.exports = {
+  plugins: [
+    autoprefixer,
+    cssnano({ preset: 'default' })
+  ]
+};
