+++ src/index.html
@@ -9,13 +9,13 @@
     <meta name="keywords" content="Mesto, Россия, фотографии">
     <meta name="author" content="дизайн - Яндекс.Практикум">
     <meta name="author" content="верстка - Бобрик Александр">
-    <link rel="stylesheet" href="./pages/index.css">
+    <!--<link rel="stylesheet" href="./pages/index.css">-->
 </head>
 <body class="page">
 
   <!--шапка сайта-->
   <header class="header">
-    <img src="./images/logo.svg" alt="логотип Mesto" class="logo">
+    <img src="<%=require('./images/logo.svg')%>" alt="логотип Mesto" class="logo">
   </header>
 
   <!--основной контент страницы-->
@@ -23,13 +23,13 @@
 
     <!--профиль пользователя-->
     <section class="profile">
-      <img src="./images/avatar.jpg" alt="фотография Жак-Ив Кусто" class="profile__avatar">
+      <img src="<%=require('./images/avatar.jpg')%>" alt="фотография Жак-Ив Кусто" class="profile__avatar">
       <div class="profile__information">
         <h1 class="profile__title">Жак-Ив Кусто</h1>
         <p class="profile__subtitle">Исследователь океана</p>
       </div>
-      <button type="button" class="button button_function_edit-profile profile__edit-button"><img src="./images/icon-pen.svg" alt="иконка редактирования в виде ручки" class="button__icon-pen"></button>
-      <button type="button" class="button profile__add-button"><img src="./images/icon-plus.svg" alt="иконка добавления в виде плюса" class="button__icon-plus"></button>
+      <button type="button" class="button button_function_edit-profile profile__edit-button"><img src="<%=require('./images/icon-pen.svg')%>" alt="иконка редактирования в виде ручки" class="button__icon-pen"></button>
+      <button type="button" class="button profile__add-button"><img src="<%=require('./images/icon-plus.svg')%>" alt="иконка добавления в виде плюса" class="button__icon-plus"></button>
     </section>
 
     <!--галерея фотографий-->
@@ -47,7 +47,7 @@
   <!--модальное окно редактирования профиля-->
   <div class="popup popup__edit-profile">
     <div class="popup__container">
-      <button type="button" class="button button_function_close-popup popup__close-button"><img src="./images/icon-close.svg" alt="иконка закрытия окна редактирования профиля" class="button__icon-close"></button>
+      <button type="button" class="button button_function_close-popup popup__close-button"><img src="<%=require('./images/icon-close.svg')%>" alt="иконка закрытия окна редактирования профиля" class="button__icon-close"></button>
       <h2 class="popup__title">Редактировать профиль</h2>
       <form name="profile-information" action="URL" method="post" class="popup__form">
         <fieldset class="popup__set">
@@ -64,7 +64,7 @@
   <!--модальное окно добавления карточек-->
   <div class="popup popup__add-card">
     <div class="popup__container">
-      <button type="button" class="button button_function_close-popup popup__close-button"><img src="./images/icon-close.svg" alt="иконка закрытия окна редактирования профиля" class="button__icon-close"></button>
+      <button type="button" class="button button_function_close-popup popup__close-button"><img src="<%=require('./images/icon-close.svg')%>" alt="иконка закрытия окна редактирования профиля" class="button__icon-close"></button>
       <h2 class="popup__title">Новое место</h2>
       <form name="new-card" action="URL" method="post" class="popup__form">
         <fieldset class="popup__set">
@@ -81,7 +81,7 @@
   <!--модальное окно просмотра фотографий-->
   <div class="popup popup_photo">
     <div class="popup__container popup__container_photo">
-      <button type="button" class="button button_function_close-popup popup__close-button"><img src="./images/icon-close.svg" alt="иконка закрытия окна редактирования профиля" class="button__icon-close"></button>
+      <button type="button" class="button button_function_close-popup popup__close-button"><img src="<%=require('./images/icon-close.svg')%>" alt="иконка закрытия окна редактирования профиля" class="button__icon-close"></button>
       <img src="" alt="" class="popup__photo">
       <h2 class="popup__title popup__title_photo"></h2>
     </div>
@@ -92,12 +92,12 @@
     <li class="photos__item">
         <img src=" " alt=" " class="photos__photo">
         <h2 class="photos__caption"></h2>
-        <button type="button" class="button button_function_like photos__like-button"><img src="./images/like-no.svg" alt="лайк выключен" class="button__icon-like"></button>
-        <button type="button" class="button button_function_trash"><img src="./images/icon-trash.svg" alt="иконка мусорной корзины" class="button__icon-trash"></button>
+        <button type="button" class="button button_function_like photos__like-button"><img src="<%=require('./images/like-no.svg')%>" alt="лайк выключен" class="button__icon-like"></button>
+        <button type="button" class="button button_function_trash"><img src="<%=require('./images/icon-trash.svg')%>" alt="иконка мусорной корзины" class="button__icon-trash"></button>
     </li>
   </template>
 
   <!--подключение скрипта-->
-  <script type="module" src="./components/index.js"></script>
+  <!--<script type="module" src="./components/index.js"></script>-->
 </body>
 </html>
